+++ Resources/ModelArray.py
@@ -1047,3 +1047,22 @@ Bluetooth_BRCM20702_v1 = [
     "iMac13,2",
     "iMac13,3",
 ]
+
+T2_Models = [
+    "MacBookAir8,1",
+    "MacBookAir8,2",
+    "MacBookAir9,1",
+    "MacBookPro15,1",
+    "MacBookPro15,2",
+    "MacBookPro15,3",
+    "MacBookPro15,4",
+    "MacBookPro16,1",
+    "MacBookPro16,2",
+    "MacBookPro16,3",
+    "MacBookPro16,4",
+    "Macmini8,1",
+    "iMac20,1",
+    "iMac20,2",
+    "iMacPro1,1",
+    "MacPro7,1",
+]